Combining Bench File
Made a new benchmark file to combine the benches for add, sub, inverse and 5th root.

Fixed some formatting issues.

Also replaced overflowing_add_unsigned by overflowing_add as it seems to be faster for some reason.
