+++ mersenne-31/Cargo.toml
@@ -15,15 +15,7 @@ p3-util = { path = "../util" }
 rand = "0.8.5"
 
 [[bench]]
-name = "inverse"
-harness = false
-
-[[bench]]
-name = "root_5"
-harness = false
-
-[[bench]]
-name = "sub"
+name = "bench_field"
 harness = false
 
 [dev-dependencies]
